Upload cirros images as public
For the cross_tenant_traffic test, the img flavor needs
to be available in tenant, and alt_img in alt_tenant.
Since we however use img and alt_img from the same image,
the image needs to be made public so that it can be
accessed from both tenants. Alternative would be to
upload the image twice, once for each tenant and set
the img ids correspondingly.